Move read2_abbreviated function into read2.rs
